requirements.txt
git+https://github.com/adsabs/ADSMicroserviceUtils.git@master
Flask==0.12.2
Flask-RESTful==0.3.6
Flask-Testing==0.6.2
flask-discoverer==0.0.2
flask-consulate==0.1.2
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-Migrate==1.6.0
psycopg2==2.6.1
git+https://github.com/jonnybazookatone/flask-watchman@7d002bbba5babc5545f682045a2574b68908d0ce
boto3==1.1.1
boto==2.39.0
botocore==1.1.12
